refactor(frontend/controller): move down the results list mapping to result component
